test: Fix KonnectorAccounts tests
Extacted fetchTriggersFromAccounts to be able to mock it
