+++ packages/cozy-harvest-lib/src/connections/accounts.js
@@ -232,3 +232,16 @@ export const createOrUpdateAccount = async ({
 
   return await saveAccount(client, konnector, accountToSave)
 }
+
+/**
+ * Returns { trigger, account } list
+ */
+export const fetchAccountsFromTriggers = async (client, triggers) => {
+  const accountCol = client.collection('io.cozy.accounts')
+  const accountIdToTrigger = keyBy(triggers, triggersModel.getAccountId)
+  const accountIds = Object.keys(accountIdToTrigger)
+  const { data: accounts } = await accountCol.getAll(accountIds)
+  return accounts
+    .filter(Boolean)
+    .map(account => ({ account, trigger: accountIdToTrigger[account._id] }))
+}
